Create a new vue component for the selection controls.
